requirements.txt
pytest>=2.8.1
docker-py
